+++ llvm/include/llvm/MC/MCInstrItineraries.h
@@ -62,7 +62,7 @@ struct InstrStage {
   };
 
   unsigned Cycles_;  ///< Length of stage in machine cycles
-  uint64_t Units_;   ///< Choice of functional units
+  unsigned Units_;   ///< Choice of functional units
   int NextCycles_;   ///< Number of machine cycles to next stage
   ReservationKinds Kind_; ///< Kind of the FU reservation
 
@@ -72,7 +72,7 @@ struct InstrStage {
   }
 
   /// getUnits - returns the choice of FUs
-  uint64_t getUnits() const {
+  unsigned getUnits() const {
     return Units_;
   }
